// store/searchHotel/types.ts
export enum HotelFilterTypes {
  SHOW_MAP_VIEW_OVERLAY = '[Search Hotel] SHOW MAP OVERLAY',
  SHOW_SORT_BY_OVERLAY = '[Search Hotel] SHOW SORT BY OVERLAY',
  SHOW_FILTER_OVERLAY = '[Search Hotel] SHOW FILTER OVERLAY',
  CLOSE_OVERLAY = '[Search Hotel] CLOSE OVERLAY',
  CHANGE_VIEW_TYPE = '[Search Hotel] CHANGE VIEW TYPE',
  OPEN_FILTER_SECTION = '[Search Hotel] OPEN FILTER SECTION',
  GET_HOTEL_POSITION = '[Search Hotel] GET HOTEL POSITION',
  OPEN_HEADER_SEARCH_BOX = '[Search Hotel] OPEN HEADER SEARCH BOX',
  CLOSE_HEADER_SEARCH_BOX = '[Search Hotel] CLOSE HEADER SEARCH BOX',
  TOGGLE_HEADER_SEARCH_BOX = '[Search Hotel] TOGGLE HEADER SEARCH BOX'
}